+++ src/main/java/com/test/Kstream/services/ProfileProcessor.java
@@ -10,11 +10,11 @@ public interface ProfileProcessor {
     //Profile calculations
 
     BankProfilesEntity initializeProfile(String cardNumber);
-    List<Float> calculateActivityProfile(List<Float> actual,int transactionsCount, Activity activity);
-    List<Float> calculateAmountProfile(List<Float> actual, int transactionsCount, AmountOfTransaction amount);
-    List<Float> calculateAuthProfile(List<Float> actual, int transactionsCount , Authorization authorization);
-    List<Float> calculateCountryProfile(List<Float> actual, int transactionsCount, CountryOfTheCommerce country);
-    List<Float> calculateTimeProfile(List<Float> actual, int transactionsCount, TimeOfTransaction time);
-    List<Float> calculateProccProfile(List<Float> actual,int transactionsCount, ProcessCode processCode);
-    List<Float> calculateRespProfile(List<Float> actual,int transactionsCount,ResponseCode responseCode);
+    List<Float> calculateActivityProfile(List<Float> actual,float transactionsCount, Activity activity);
+    List<Float> calculateAmountProfile(List<Float> actual, float transactionsCount, AmountOfTransaction amount);
+    List<Float> calculateAuthProfile(List<Float> actual, float transactionsCount , Authorization authorization);
+    List<Float> calculateCountryProfile(List<Float> actual, float transactionsCount, CountryOfTheCommerce country);
+    List<Float> calculateTimeProfile(List<Float> actual, float transactionsCount, TimeOfTransaction time);
+    List<Float> calculateProccProfile(List<Float> actual,float transactionsCount, ProcessCode processCode);
+    List<Float> calculateRespProfile(List<Float> actual,float transactionsCount,ResponseCode responseCode);
 }
